refactor(router): remove using plugin for router
