+++ api/health.js
@@ -1,21 +1,14 @@
 'use strict';
 
-const Config = require('../config');
-
-module.exports  = {
-    name: 'healthCheck',
-    version: Config.version,
-    register: async function(server, options) {
-        server.route({
-            method: 'GET',
-            path:'/health',
-            handler: (request, h) => {
-                return {'status': 'I am still alive!'};
-            },
-            options:{
-                tags: ['api']
-            }
-        });
+module.exports  = [
+    {
+        method: 'GET',
+        path:'/health',
+        handler: (request, h) => {
+            return {'status': 'I am still alive!'};
+        },
+        options:{
+            tags: ['api']
+        }
     }
-}
-
+]
