+++ api/index.js
@@ -1,19 +1,19 @@
 'use strict';
 
 const Config = require('../config');
+const health = require('./health');
 
-module.exports = function(server) {
-    server.route({
-        method: 'GET',
-        path:'/',
-        handler: (request, h) => {
-            return {
-                "name": Config.name,
-                "version": Config.version
-            }
-        },
-        options: {
-            tags: ['api']
+const root = {
+    method: 'GET',
+    path:'/',
+    handler: (request, h) => {
+        return {
+            "name": Config.name,
+            "version": Config.version
         }
-    });
+    },
+    options: {
+        tags: ['api']
+    }
 }
+module.exports = [].concat(root, health);
